feat(tt): split apart exercise template components into individual files
