+++ apps/total-typescript/src/lib/exercises.ts
@@ -10,7 +10,7 @@ export const ExerciseSchema = z.object({
   title: z.string(),
   slug: z.string(),
   description: z.nullable(z.string()).optional(),
-  body: z.any().array().optional(),
+  body: z.nullable(z.any().array().optional()),
   stackblitz: z.nullable(z.string()).optional(),
   muxPlaybackId: z.nullable(z.string()).optional(),
   transcript: z.nullable(z.any().array()).optional(),
@@ -76,32 +76,20 @@ export const getExercise = async (
       title,
       description,
       "slug": slug.current,
-      ${
-        includeMedia
-          ? `      
         body,
         "stackblitz": resources[@._type == 'stackblitz'][0].openFile,
         "muxPlaybackId": resources[@->._type == 'videoResource'][0]-> muxAsset.muxPlaybackId,
         "transcript": resources[@->._type == 'videoResource'][0]-> castingwords.transcript,
-      `
-          : ''
-      }
       "solution": resources[@._type == 'solution'][0]{
         _key,
         _type,
         "_updatedAt": ^._updatedAt,
         title,
         description,
-        ${
-          includeMedia
-            ? `      
-          body,
-          "stackblitz": resources[@._type == 'stackblitz'][0].openFile,
-          "muxPlaybackId": resources[@->._type == 'videoResource'][0]-> muxAsset.muxPlaybackId,
-          "transcript": resources[@->._type == 'videoResource'][0]-> castingwords.transcript,
-        `
-            : ''
-        }
+        body,
+        "stackblitz": resources[@._type == 'stackblitz'][0].openFile,
+        "muxPlaybackId": resources[@->._type == 'videoResource'][0]-> muxAsset.muxPlaybackId,
+        "transcript": resources[@->._type == 'videoResource'][0]-> castingwords.transcript,
         "slug": slug.current,
       }
     }`
