--- apps/total-typescript/src/pages/workshops/lesson-paywall.tsx
@@ -1,7 +1,0 @@
-import * as React from 'react'
-
-const LessonPaywall: React.FC = () => {
-  return <div>LessonPaywall</div>
-}
-
-export default LessonPaywall
