+++ apps/total-typescript/src/server/rewrite-next-response-to-path.ts
@@ -1,7 +1,5 @@
 import {NextRequest, NextResponse} from 'next/server'
 
 export function rewriteToPath(path: string, req: NextRequest) {
-  const baseUrlForRewrites = req.nextUrl.clone()
-  baseUrlForRewrites.pathname = path
-  return NextResponse.rewrite(baseUrlForRewrites)
+  return NextResponse.rewrite(new URL(path, req.url))
 }
